Fix push notification handeling
