+++ src/navigation/AppNavigator.tsx
@@ -3,11 +3,9 @@ import {createDrawerNavigator} from '@react-navigation/drawer';
 import {NavigationContainer} from '@react-navigation/native';
 import {createStackNavigator} from '@react-navigation/stack';
 import {Icon, TopNavigationAction} from '@ui-kitten/components';
-import {ChainApiContext} from 'context/ChainApiContext';
-import LoadingView from 'presentational/LoadingView';
-import React, {useContext} from 'react';
+import React from 'react';
 import {AddAccountScreen} from 'screen/AddAccountScreen/AddAccountScreen';
-import {ApiLoadingScreen} from 'screen/ApiLoadingScreen';
+import {DeeplinkNavigatorScreen} from 'screen/DeeplinkNavigatorScreen';
 import {BalanceScreen} from 'screen/BalanceScreen';
 import {CouncilScreen} from 'screen/Council/CouncilScreen';
 import {MotionsScreen} from 'screen/Council/MotionsScreen';
@@ -138,18 +136,13 @@ function ApiLoadedNavigator() {
 const AppStack = createStackNavigator<AppStackParamList>();
 
 function AppNavigator() {
-  const {api} = useContext(ChainApiContext);
   const {theme} = useTheme();
 
   // Start: app start hooks
   useTurnOnAllNotificationsOnAppStartForAndroid();
   // End: app start hooks
 
-  const {pushAuthorizationStatus, isLoading} = usePushAuthorizationStatus();
-
-  if (isLoading) {
-    return <LoadingView />;
-  }
+  const {pushAuthorizationStatus} = usePushAuthorizationStatus();
 
   return (
     <NavigationContainer linking={routeKeys.linking} theme={theme === 'dark' ? darkTheme : lightTheme}>
@@ -157,8 +150,8 @@ function AppNavigator() {
         {pushAuthorizationStatus && pushAuthorizationStatus === messaging.AuthorizationStatus.NOT_DETERMINED ? (
           <AppStack.Screen name={routeKeys.permissionGrantingPromptScreen} component={PermissionGrantingPrompt} />
         ) : undefined}
-        {api ? <AppStack.Screen name={routeKeys.apiLoadedNavigatorScreen} component={ApiLoadedNavigator} /> : undefined}
-        <AppStack.Screen name={routeKeys.apiLoadingScreen} component={ApiLoadingScreen} />
+        <AppStack.Screen name={routeKeys.apiLoadedNavigatorScreen} component={ApiLoadedNavigator} />
+        <AppStack.Screen name={routeKeys.deeplinkNavigatorScreen} component={DeeplinkNavigatorScreen} />
       </AppStack.Navigator>
     </NavigationContainer>
   );
