+++ src/navigation/routeKeys.ts
@@ -6,7 +6,7 @@ export const apiLoadingNavigatorScreen = 'ApiLoadingNavigator' as const;
 export const permissionGrantingPromptScreen = 'PermissionsGrantingPrompt' as const;
 export const dashboardNavigator = 'DashboardNavigator' as const;
 
-export const apiLoadingScreen = 'Api Loading' as const;
+export const deeplinkNavigatorScreen = 'DeeplinkNavigatorScreen' as const;
 export const addAccountScreen = 'AddAccountScreen' as const;
 export const balanceScreen = 'BalanceScreen' as const;
 
@@ -33,7 +33,7 @@ export const linking: LinkingOptions = {
   config: {
     initialRouteName: apiLoadedNavigatorScreen,
     screens: {
-      [apiLoadingScreen]: 'api/:network/:redirectTo?',
+      [deeplinkNavigatorScreen]: 'api/:network/:redirectTo?',
       [apiLoadedNavigatorScreen]: {
         initialRouteName: dashboardNavigator,
         path: '',
